backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.6.0
sqlalchemy>=2.0.25
pandas>=2.2.0
python-dotenv>=1.0.1
psycopg2-binary>=2.9.9
httpx>=0.26.0
plaid-python>=4.0.0
python-jose[cryptography]
python-multipart
bcrypt
authlib
itsdangerous
google-auth
